refactor(funput-term): group modules into folders; keep files <150 LOC
Reorganize the flat src/ into four responsibility-scoped folders so each
file's non-test code stays under 150 lines:

- runtime/  (was app.rs + plumbing): mod = run() orchestration, driver =
  forward_input + Status, resize, input/inject/output/state
- terminal/ (was term.rs): mod = RawModeGuard, console = Windows VT/codepage
  (#[cfg(windows)]), indicator = Mux/title/cursor OSC
- config/   mod + schema (serde FileSettings) + parse (env overlay)
- install/  mod (Shell + run) + snippet (pure builder)

Public path funput_term::app -> funput_term::runtime (the sole external
consumer, funput-cli, is updated). run() now rejects an empty command
instead of index-panicking on a public entry point.

Tests kept inline (they exercise private internals). 48 tests pass; clippy
-D warnings clean on native and x86_64-pc-windows-msvc.
