+++ crates/funput-term/src/runtime/inject.rs
