+++ crates/funput-term/src/runtime/input.rs
